Revert "V.2 navbar"
