+++ scss/style.scss
@@ -1,137 +1,5 @@
-// Navigation bar stying 
-#Navbar{
-    padding: 0 100px;
-    height: 70px;
-    display: grid;
-    grid-template-columns: 2fr 1fr;
-    #NavBar_leftside{
-        display: grid;
-        grid-template-columns: 3em auto;
-        #navbar_menubtn{
-            border: none;
-            font-weight: bolder;
-        }
-        #name_form{
-            display: flex;
-            #navbar_search_form{
-                min-width: 30em;
-                padding-top: 1em;
-                #search_NavbarFormInput{
-                    border-radius: 0.25rem 0 0 0.25rem;
-                }
-                #search_NavbarFormBtn{
-                    border-radius: 0 0.25rem 0.25rem 0;
-                    margin-left: -10px;
-                    background-color: #E8F0F4;
-                    border-color: transparent;
-                    color: #003A56;
-                }
-            }
-        }
-    }
-    #NavBar_rightside{
-        div{
-            width:fit-content;
-            float: right;
-            .active{
-                background-size: 100%;
-                background-color: #003A56 ;
-                color:#E8F0F4;
-            }
-        }
-    }
-}
-.navbar_image{
-    margin: 3px 0 0 10px;
-    width: 190px;
-    height: 50px;
-    object-fit: cover;
-}
-.navbar_offcanvas{
-    margin-right: 10px;
-}
-@keyframes NavLinkHover {
-    from {
-        background-size:0;
-        background-color: transparent;
-        color: white;
-    }
-    to {
-        background-size: 100%;
-        background-color: #E8F0F4;
-        color: #003A56;
-    }
-}
-@keyframes username_dropdown_hover {
-    from {
-        background-size:0;
-        background-color: transparent;
-        color: #003A56;
-    }
-    to {
-        background-size: 100%;
-        background-color: #003A56 ;
-        color:#E8F0F4;
-    }
-}
-.username_Dropdown_items{
-    font-weight: bolder;
-    color: #003A56;
-
-}
-.username_Dropdown_items:hover{
-    animation-name: username_dropdown_hover;
-    animation-duration: 2s;
-    animation-fill-mode: forwards;
-}
-.Navitem:hover{
-    background-color: transparent;
-    animation-name: NavLinkHover;
-    animation-fill-mode: forwards;
-    animation-duration: 2s;
-}
-#username_Dropdown::after{
-    content: none;
-}
-#username_Dropdown{
-    color: white;
-    font-weight: bold;
-    font-size: larger;
-    a{
-        color: white;
-        font-weight: bold;
-        font-size: larger;
-    }
-}
-
-.NavbarLink, #offcanvasNavbarDropdown-expand-true{
-    color: white;
-    font-weight: bold;
-    font-size: larger;
-    
-    .dropdown-menu{
-        background-color: transparent;
-        border: none;
-    }
-    .dropdown-item{
-        margin-left:0.5rem;
-        padding-left: 5px;
-    }
-    .dropdown-menu:hover{
-        background-color: transparent;
-        border-bottom: none;
-    }
-    
-}
-.NavFonts{
-    font-family: "Nunito"!important;
-}
-// ######################################################
 .body{
-    margin-top: -16px;
-    padding-top: 3.9em;
-    font-family: "Nunito"!important;
-    background-color: E8F0F4;
+    margin-top: 3.9em;
     .FormPage{
         height: 50%;
         min-height: 27em;
