+++ server.js
@@ -324,5 +324,5 @@ server.use(express.static('public')); // use this middleware before get method.
 
 var server_port = process.env.YOUR_PORT || process.env.PORT || 3000;
 server.listen(server_port, async ()=>{
-    console.log("Server is listening on port 4200");
+    console.log("Server is listening on port 3000");
 });
