+++ src/components/Navigationbar/Navigationbar.js
@@ -1,30 +1,9 @@
 import { NavLink } from "react-router-dom";
-import { Nav } from "react-bootstrap";
+import { Container, Nav } from "react-bootstrap";
 import Navbar from "react-bootstrap/Navbar";
-import Button from 'react-bootstrap/Button';
-import Form from 'react-bootstrap/Form';
 import NavDropdown from 'react-bootstrap/NavDropdown';
-import Offcanvas from 'react-bootstrap/Offcanvas';
-import images from './images.json';
-import { useState, useEffect } from "react";
-
-
 
 var Navigationbar = ()=>{
-    const [show, setShow] = useState(false);
-    const [categoryShow, setCategoryShow] = useState(false);
-    var site_logo = images.name_logo;
-    var theme_color= "#003A56";
-    var username = sessionStorage.getItem("username");
-    var returnNav;
-    var rightSide;
-
-    useEffect(()=>{
-        var path = window.location.pathname;
-        if(path == "/" || path == "/myItems" || path == "/myCart"){setCategoryShow(true)}
-        else{setCategoryShow(false)}
-    }, [window.location.pathname]);
-
     var onDropDown = (e)=>{
         var path = window.location.pathname;
         var pathArr = window.location.pathname.split("/");
@@ -40,99 +19,58 @@ var Navigationbar = ()=>{
         sessionStorage.removeItem("username");
         window.location.href = "/";
     }
-
-    var showChange = ()=>setShow(!show);
-    
-    // display the Navlinks as per sign in status
+    let username = (sessionStorage.getItem("username"));
+    var returnNav;
+    var userDisplay;
     if(username === null){
-        returnNav = <Nav id="NavbarBody" className="justify-content-end flex-grow-1 pe-3">
-            <Nav.Link className="NavFonts NavbarLink Navitem" as={NavLink} to="/" onClick={showChange}>Home</Nav.Link>
-            <Nav.Link className="NavFonts NavbarLink Navitem" as={NavLink} to="/register" onClick={showChange} >Register</Nav.Link>
-            <Nav.Link className="NavFonts NavbarLink Navitem"  as={NavLink} to="/signin" onClick={showChange}>Sign In</Nav.Link>
-            {categoryShow == true ? 
-            <NavDropdown title="Category" className="NavbarLink NavFonts" onSelect={onDropDown}             id='offcanvasNavbarDropdown-expand-true' style={{fontSize: "large"}}>
-                <NavDropdown.Item  className="NavFonts NavbarLink Navitem" eventKey="All">All</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Toys">Toys</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Home">Home</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Fashion">Fashion</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Electronics">Electronics</NavDropdown.Item>
-                <NavDropdown.Divider style={{color: "white"}}/>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Lowtohigh">Price: Low to High</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Hightolow">Price: Hign to Low</NavDropdown.Item>
-            </NavDropdown> : <></>}
-        </Nav>;
-        rightSide = <></>
+        returnNav = <Nav id="Navbarclass" className="me-auto">
+                        <Nav.Link as={NavLink} to="/register">Register</Nav.Link>
+                        <Nav.Link as={NavLink} to="/signin">Sign In</Nav.Link>
+                        <NavDropdown title="Category" onSelect={onDropDown}>
+                            <NavDropdown.Item eventKey="All">All</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Toys">Toys</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Home">Home</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Fashion">Fashion</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Electronics">Electronics</NavDropdown.Item>
+                            <NavDropdown.Divider />
+                            <NavDropdown.Item eventKey="Lowtohigh">Price: Low to High</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Hightolow">Price: Hign to Low</NavDropdown.Item>
+                        </NavDropdown>
+                    </Nav>;
+        userDisplay = <Navbar.Text className="justify-content-end">Not Signed In</Navbar.Text>;
     }else{
-        returnNav = <Nav id="NavbarBody" className="justify-content-end flex-grow-1 pe-3">
-            <Nav.Link className="NavFonts NavbarLink Navitem" as={NavLink} to="/" onClick={showChange}>Home</Nav.Link>
-            <Nav.Link className="NavFonts NavbarLink Navitem" as={NavLink} to="/myItems" onClick={showChange}>My Item</Nav.Link>
-            <Nav.Link className="NavFonts NavbarLink Navitem" as={NavLink} to="/addItem" onClick={showChange}>Add Item</Nav.Link>
-            {/* <Nav.Link as={NavLink} to="#" onClick={logout}>Log Out</Nav.Link> */}
-            {categoryShow == true ? 
-            <NavDropdown title="Category" className="NavbarLink NavFonts" onSelect={onDropDown}             id='offcanvasNavbarDropdown-expand-true' style={{fontSize: "large"}}>
-                <NavDropdown.Item  className="NavFonts NavbarLink Navitem" eventKey="All">All</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Toys">Toys</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Home">Home</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Fashion">Fashion</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Electronics">Electronics</NavDropdown.Item>
-                <NavDropdown.Divider style={{color: "white"}}/>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Lowtohigh">Price: Low to High</NavDropdown.Item>
-                <NavDropdown.Item className="NavFonts NavbarLink Navitem" eventKey="Hightolow">Price: Hign to Low</NavDropdown.Item>
-            </NavDropdown> : <></>}
-        </Nav>
-        rightSide=<NavDropdown id="username_Dropdown" title={username} className="NavFonts">
-            <NavDropdown.Item className="username_Dropdown_items NavFonts"
-            id="logout" onClick={logout} >Log Out</NavDropdown.Item>
-            <NavDropdown.Item className="username_Dropdown_items NavFonts" id="myCart" as={NavLink} to="/myCart">My Cart</NavDropdown.Item>
-        </NavDropdown>
+        returnNav = <Nav id="Navbarclass" className="me-auto">
+                        <Nav.Link as={NavLink} to="/myCart">My Cart</Nav.Link>
+                        <Nav.Link as={NavLink} to="/myItems">My Item</Nav.Link>
+                        <Nav.Link as={NavLink} to="/addItem">Add Item</Nav.Link>
+                        <Nav.Link as={NavLink} to="#" onClick={logout}>Log Out</Nav.Link>
+                        <NavDropdown title="Category" onSelect={onDropDown}>
+                            <NavDropdown.Item eventKey="All">All</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Toys">Toys</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Home">Home</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Fashion">Fashion</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Electronics">Electronics</NavDropdown.Item>
+                            <NavDropdown.Divider />
+                            <NavDropdown.Item eventKey="Lowtohigh">Price: Low to High</NavDropdown.Item>
+                            <NavDropdown.Item eventKey="Hightolow">Price: Hign to Low</NavDropdown.Item>
+                        </NavDropdown>
+                    </Nav>;
+        userDisplay = <Navbar.Text className="justify-content-end">Signed in as <strong>{username}</strong></Navbar.Text>;
     }
-
     return (
-        <>
-            <Navbar variant="dark" id="Navbar" expand='true' className="mb-3" style={{backgroundColor: theme_color, fontFamily: "Nunito"}}>
-                <div id="NavBar_leftside">
-                    {/* menu button */}
-                    <Navbar.Toggle aria-controls={`offcanvasNavbar-expand-true`} id="navbar_menubtn" onClick={showChange}/>
-                    {/* Off Canvas  */}
-                    <Navbar.Offcanvas show={show}
-                    id={`offcanvasNavbar-expand-true`}
-                    aria-labelledby={`offcanvasNavbarLabel-expand-true`}
-                    style={{backgroundColor: theme_color}} onHide={showChange}
-                    >
-                        <Offcanvas.Header closeButton closeVariant="white" className="navbar_offcanvas" >
-                        <Offcanvas.Title id={`offcanvasNavbarLabel-expand-true`}>
-                            {/* add the my store logo */}
-                            <Navbar.Brand as={NavLink} to="/"><img className = "navbar_image" src={site_logo} /></Navbar.Brand>
-                        </Offcanvas.Title>
-                    </Offcanvas.Header>
-                    {/* this is body of side screen */}
-                    <Offcanvas.Body>
+        <Navbar fixed="top" style={{backgroundColor: "white"}}>
+            <Container>
+                <Navbar.Brand as={NavLink} to="/" style={{fontWeight: "bold"}}>My Store</Navbar.Brand>
+                    <Navbar.Toggle aria-controls="basic-navbar-nav" />
+                    <Navbar.Collapse id="basic-navbar-nav">
                         {returnNav}
-                    </Offcanvas.Body>
-                    </Navbar.Offcanvas>
-                    {/* add the my store logo */}
-                    <div id="name_form">
-                        <Navbar.Brand as={NavLink} to="/"><img className = "navbar_image"   src={site_logo} /></Navbar.Brand>
-                        <Form className="d-flex" id="navbar_search_form">
-                            <Form.Control
-                                type="search"
-                                placeholder="Search"
-                                className="me-2"
-                                aria-label="Search"
-                                id="search_NavbarFormInput"
-                            />
-                            <Button className="btn" id="search_NavbarFormBtn"><i className="fa fa-search"></i></Button>
-                        </Form>
-                    </div>
-                </div>
-                <div id="NavBar_rightside">
-                    <div>
-                        {rightSide}
-                    </div>
-                </div>
-            </Navbar>
-        </>
+                        <Nav>
+                            {userDisplay}
+                        </Nav>
+                    </Navbar.Collapse>
+            </Container>
+        </Navbar>
     );
-};
+}
 
 export default Navigationbar;
