+++ views/header.ejs
@@ -28,9 +28,6 @@
       sizes="16x16"
       href="./bag.ico"
     />
-        
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-    <link href='https://fonts.googleapis.com/css?family=Nunito' rel='stylesheet'>
     <title>My Store</title>
 </head>
 <body>
